requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyyaml==6.0.1
matplotlib==3.8.2
numpy==1.26.2
gunicorn==21.2.0
requests==2.31.0
